backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
python-dotenv>=1.0
django-cors-headers>=4.3
requests
gunicorn
cryptography>=42.0.0